v1.4.0: blueprint decomposition, SQLite, multi-user auth (#31)
* Fix: ruff format main.py (vendor list one-per-line)


* refactor: decompose web.py into Flask blueprints

Splits ~2,000-line web.py into 7 blueprints + shared helpers.
Zero behavior change — all routes respond identically.

- web.py: app factory only (184 lines, down from ~2,000)
- blueprints/auth.py: /login, /logout (seed for future LDAP/OIDC)
- blueprints/audit.py: /audit, /bulk_audit, /diff, /connect
- blueprints/history.py: /archive/*, /activity/*
- blueprints/schedules.py: /schedules/*
- blueprints/settings_bp.py: /settings/*, /license/*
- blueprints/reports.py: /reports/*
- blueprints/api_v1.py: /api/v1/*
- _vendor_helpers.py: detect_vendor(), validate_vendor_format(), extract_hostname()
- _helpers.py: _require_auth_impl(), _err(), _make_temp_path()

All 181 tests pass. ruff clean.


* feat: migrate collection stores to SQLite (cashel.db)

Replaces per-file JSON storage in archive, activity_log, and
schedule_store with a single SQLite database (cashel.db).

- db.py: thread-local connections, WAL mode, idempotent schema init,
  JSON→SQLite auto-migration on first startup
- archive.py: audits table; auto-versioning via SELECT MAX(version)
  in transaction; fcntl lock removed
- activity_log.py: activity table; bool↔int, dict↔JSON on read/write
- schedule_store.py: schedules table; bool fields as 0/1 int;
  encryption helpers unchanged
- web.py: init_db() called on startup after blueprint registration
- tests/test_db.py: 21 new tests (202 total passing)

Public API of all three modules is unchanged. Original JSON files
left in place for rollback safety.


* feat: multi-user auth Phase 1 — local username/password

Replaces single shared API key with per-user authentication.

- user_store.py: CRUD for users table; password hashing via
  werkzeug.security; per-user Fernet-encrypted API keys;
  UserValidationError for bad input; last-admin guard on delete
- db.py: users table added to schema (id, username, password_hash,
  role, api_key_enc, created_at)
- _helpers.py: _require_auth_impl updated for multi-user (user_id in
  session, g.current_user set on every request); _require_role
  decorator added for blueprint-level role enforcement
- blueprints/auth.py: /login, /logout, /setup (first-run wizard),
  /auth/users CRUD, /auth/change-password, /auth/generate-api-key,
  /auth/revoke-api-key; in-memory lockout after 5 failures
- blueprints/audit.py: @_require_role("admin","auditor") on /audit,
  /bulk_audit, /connect
- blueprints/schedules.py: @_require_role on write routes
- blueprints/settings_bp.py: @_require_role("admin") on write routes
- templates/login.html: username + password fields (API key field removed)
- templates/setup.html: first-run admin account creation wizard
- templates/index.html: user management table, per-user API key
  controls, change-password — all in Auth settings pane
- settings.py: api_key_enc and save_api_key() removed
- web.py: passes current_user to index template
- tests/test_auth.py: 32 new tests (234 total passing)

Roles: admin (full), auditor (audit+schedules+export), viewer (read-only).
API key header auth (X-API-Key) preserved for CI/CD — per user.
First-run /setup wizard auto-enables auth on completion.


* chore: bump version to 1.4.0

Covers web.py blueprint decomposition, SQLite migration,
and multi-user auth Phase 1.


---------
