+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "cashel"
-version = "1.3.0"
+version = "1.4.0"
 description = "Firewall configuration auditing tool"
 authors = [
     {name = "Your Name", email = "you@example.com"}
