+++ src/cashel/settings.py
@@ -86,10 +86,6 @@ def get_settings() -> dict:
         # Decrypt smtp_password — stored encrypted, exposed in-process as plaintext
         if data.get("smtp_password_enc"):
             merged["smtp_password"] = decrypt(data["smtp_password_enc"])
-        # Decrypt api_key — stored encrypted, never exposed to the template directly
-        merged["api_key"] = (
-            decrypt(data["api_key_enc"]) if data.get("api_key_enc") else ""
-        )
         return merged
     except (FileNotFoundError, json.JSONDecodeError):
         return dict(DEFAULTS)
@@ -125,20 +121,6 @@ def save_settings(data: dict) -> dict:
         merged["smtp_password_enc"] = ""
     merged.pop("smtp_password", None)
 
-    # api_key is managed separately via /settings/generate-api-key; never passed
-    # through save_settings (it would be overwritten to empty on every settings save).
-    # Remove it from the dict — the encrypted api_key_enc is preserved from disk.
-    merged.pop("api_key", None)
-
-    # Preserve existing api_key_enc from disk if not being explicitly updated
-    try:
-        with open(SETTINGS_FILE) as _f:
-            _existing = json.load(_f)
-        if _existing.get("api_key_enc") and "api_key_enc" not in merged:
-            merged["api_key_enc"] = _existing["api_key_enc"]
-    except (FileNotFoundError, json.JSONDecodeError):
-        pass
-
     os.makedirs(os.path.dirname(SETTINGS_FILE), exist_ok=True)
     with open(SETTINGS_FILE, "w") as f:
         json.dump(merged, f, indent=2)
@@ -148,17 +130,3 @@ def save_settings(data: dict) -> dict:
     return merged
 
 
-def save_api_key(plaintext_key: str) -> None:
-    """Encrypt and persist the API key. Separate from save_settings so it is
-    never accidentally overwritten by a settings form submission."""
-    from .crypto import encrypt as _enc
-
-    try:
-        with open(SETTINGS_FILE) as f:
-            data = json.load(f)
-    except (FileNotFoundError, json.JSONDecodeError):
-        data = {}
-    data["api_key_enc"] = _enc(plaintext_key)
-    os.makedirs(os.path.dirname(SETTINGS_FILE), exist_ok=True)
-    with open(SETTINGS_FILE, "w") as f:
-        json.dump(data, f, indent=2)
